qubofinalprotype/qubofinalprotype/qubofinalprotype/requirements.txt
numpy
matplotlib
pytest
click
PyQt5
google-generativeai  # Optional: needed for Gemini AI chat & suggestions
